+++ kartothek/io_components/utils.py
@@ -2,10 +2,10 @@
 This module is a collection of helper functions
 """
 
-
 import collections
 import inspect
 import logging
+from typing import Callable, Optional
 
 import decorator
 import pandas as pd
@@ -378,7 +378,7 @@ class NoPickleFactory:
         raise TypeError("Serialization not allowed. Please use a proper store factory.")
 
 
-def _make_callable(obj):
+def _make_callable(obj) -> Optional[Callable]:
     """
     Converting a plain store to a simple callable is perfectly fine for
     eager/iter/etc backends. Must not be used if the store/factory is serialized.